+++ products/views.py
@@ -146,7 +146,12 @@ def edit_product(request, product_id):
 
 def delete_product(request, product_id):
     """ Delete a product from the store """
-    product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+
+    if request.user.is_staff or request.user.is_superuser:
+        product = get_object_or_404(Product, pk=product_id)
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
+    else:
+        messages.error(request, "You are not allowed in this area.")
+        return redirect(reverse('home'))
